pyproject.toml
[project]
name = "community-bots"
version = "0.1.0"
description = "Community Bots - PyLadies and R-Ladies Bots"
authors = [
    {name = "cosimameyer", email = "cosima.meyer@gmail.com"},
]
dependencies = [
    "beautifulsoup4==4.11.2",
    "blurhash==1.1.4",
    "certifi==2022.12.7",
    "charset-normalizer==3.0.1",
    "decorator==5.1.1",
    "feedparser==6.0.10",
    "idna==3.4",
    "lxml==4.9.2",
    "Mastodon-py==1.8.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "python-magic==0.4.27",
    "requests==2.31.0",
    "sgmllib3k==1.0.0",
    "six==1.16.0",
    "soupsieve==2.4",
    "urllib3==1.26.14",
    "atproto==0.0.55",
    "google-generativeai==0.8.3",
    "pre-commit>=4.2.0",
    "pylint>=3.3.6",
]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "CC0 1.0"}


[tool.pdm]
distribution = false
